Modified ResponsiveHTML: SESSION(), update()

+++ src/main/java/supersql/codegenerator/Mobile_HTML5/Mobile_HTML5Manager.java
@@ -243,10 +243,14 @@ public class Mobile_HTML5Manager extends Manager{
 //					String fn = ".htaccess";
 //					if(!html_env.getFileParent().isEmpty())
 //						fn = html_env.getFileParent()+GlobalEnv.OS_FS+fn;
-					PrintWriter pw = new PrintWriter(new BufferedWriter(new OutputStreamWriter(
-							new FileOutputStream(fn), GlobalEnv.DEFAULT_CHARACTER_CODE)));
-					pw.println("AddType application/x-httpd-php .html");
-					pw.close();
+					if (new File(fn).exists()){
+					    //System.out.println(".htaccess already exists");
+					}else{
+						PrintWriter pw = new PrintWriter(new BufferedWriter(new OutputStreamWriter(
+								new FileOutputStream(fn), GlobalEnv.DEFAULT_CHARACTER_CODE)));
+						pw.println("AddType application/x-httpd-php .html");
+						pw.close();
+					}
 
 					Mobile_HTML5Env.initAllFormFlg();
 					Jscss.process();	//goto 20141209
